+++ src/pages/Recruting/step3/DocumentEvaluation.tsx
@@ -1,6 +1,6 @@
 //3 - 리크루팅 : 서류 평가하기 단계 (페이지)
 import DocumentReviewPrepContainer from "../../../components/recruting/_03_document_evaluation/_01/DocumentReviewPrepContainer";
-import TopSection from "../../../components/recruting/_03_document_evaluation/TopSection";
+import TopSection from "../../../components/recruting/_03_document_evaluation/common/TopSection";
 //임시로 넣어놨어요
 import CreateApplicationFormContainer from "../../../components/recruting/_02_prepare/_05/CreateApplicationFormContainer";
